refactor: cleanup router tests, split into smaller chunks
